+++ packages/@tinacms/api-git/src/router.ts
@@ -29,6 +29,10 @@ import { show } from './show'
 import { getGitSSHUrl, isSSHUrl } from './utils/gitUrl'
 import atob from 'atob'
 
+// Don't return full error message to client incase confidential details leak
+const GIT_ERROR_MESSAGE =
+  'Git Operation failed: Check the logs for more details'
+
 export interface GitRouterConfig {
   pathToRepo?: string
   pathToContent?: string
@@ -135,8 +139,8 @@ export function router(config: GitRouterConfig = {}) {
 
     try {
       deleteFile(fileAbsolutePath)
-    } catch (e) {
-      res.status(500).json({ status: 'error', message: e.message })
+    } catch {
+      res.status(500).json({ status: 'error', message: GIT_ERROR_MESSAGE })
     }
 
     commit({
@@ -150,8 +154,8 @@ export function router(config: GitRouterConfig = {}) {
       .then(() => {
         res.json({ status: 'success' })
       })
-      .catch(e => {
-        res.status(500).json({ status: 'error', message: e.message })
+      .catch(() => {
+        res.status(500).json({ status: 'error', message: GIT_ERROR_MESSAGE })
       })
   })
 
@@ -175,8 +179,8 @@ export function router(config: GitRouterConfig = {}) {
         )
       }
       res.json({ content: req.body.content })
-    } catch (e) {
-      res.status(500).json({ status: 'error', message: e.message })
+    } catch {
+      res.status(500).json({ status: 'error', message: GIT_ERROR_MESSAGE })
     }
   })
 
@@ -193,8 +197,8 @@ export function router(config: GitRouterConfig = {}) {
         if (err) console.error(err)
       })
       res.send(req.file)
-    } catch (e) {
-      res.status(500).json({ status: 'error', message: e.message })
+    } catch {
+      res.status(500).json({ status: 'error', message: GIT_ERROR_MESSAGE })
     }
   })
 
@@ -217,10 +221,10 @@ export function router(config: GitRouterConfig = {}) {
       })
 
       res.json({ status: 'success' })
-    } catch (e) {
+    } catch {
       // TODO: More intelligently respond
       res.status(412)
-      res.json({ status: 'failure', error: e.message })
+      res.json({ status: 'failure', error: GIT_ERROR_MESSAGE })
     }
   })
 
@@ -228,10 +232,10 @@ export function router(config: GitRouterConfig = {}) {
     try {
       await openRepo(REPO_ABSOLUTE_PATH).push()
       res.json({ status: 'success' })
-    } catch (e) {
+    } catch {
       // TODO: More intelligently respond
       res.status(412)
-      res.json({ status: 'failure', error: e.message })
+      res.json({ status: 'failure', error: GIT_ERROR_MESSAGE })
     }
   })
 
@@ -246,9 +250,9 @@ export function router(config: GitRouterConfig = {}) {
       .then(() => {
         res.json({ status: 'success' })
       })
-      .catch((e: any) => {
+      .catch(() => {
         res.status(412)
-        res.json({ status: 'failure', error: e.message })
+        res.json({ status: 'failure', error: GIT_ERROR_MESSAGE })
       })
   })
 
@@ -256,10 +260,10 @@ export function router(config: GitRouterConfig = {}) {
     try {
       const summary = await openRepo(REPO_ABSOLUTE_PATH).branchLocal()
       res.send({ status: 'success', branch: summary.branches[summary.current] })
-    } catch (e) {
+    } catch {
       // TODO: More intelligently respond
       res.status(500)
-      res.json({ status: 'failure', message: e.message })
+      res.json({ status: 'failure', message: GIT_ERROR_MESSAGE })
     }
   })
 
@@ -267,10 +271,10 @@ export function router(config: GitRouterConfig = {}) {
     try {
       const summary = await openRepo(REPO_ABSOLUTE_PATH).branchLocal()
       res.send({ status: 'success', branches: summary.all })
-    } catch (e) {
+    } catch {
       // TODO: More intelligently respond
       res.status(500)
-      res.json({ status: 'failure', message: e.message })
+      res.json({ status: 'failure', message: GIT_ERROR_MESSAGE })
     }
   })
 
@@ -289,16 +293,16 @@ export function router(config: GitRouterConfig = {}) {
       }
 
       res.send({ status: 'success', branch })
-    } catch (e) {
+    } catch {
       // TODO: More intelligently respond
       res.status(500)
-      res.json({ status: 'failure', message: e.message })
+      res.json({ status: 'failure', message: GIT_ERROR_MESSAGE })
     }
   })
 
   router.get('/show/:fileRelativePath', async (req, res) => {
     try {
-      const fileRelativePath = path
+      const fileRelativePath = path.posix
         .join(CONTENT_REL_PATH, req.params.fileRelativePath)
         .replace(/^\/*/, '')
 
@@ -312,11 +316,11 @@ export function router(config: GitRouterConfig = {}) {
         content,
         status: 'success',
       })
-    } catch (e) {
+    } catch {
       res.status(501)
       res.json({
         status: 'failure',
-        message: e.message,
+        message: GIT_ERROR_MESSAGE,
         fileRelativePath: req.params.fileRelativePath,
       })
     }