+++ packages/demo-gatsby/src/components/bio.js
@@ -26,7 +26,7 @@ limitations under the License.
 import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
 import Image from "gatsby-image"
-import { useJsonForm } from "gatsby-tinacms-json"
+import { useLocalJsonForm } from "gatsby-tinacms-json"
 
 import { rhythm } from "../utils/typography"
 
@@ -74,7 +74,7 @@ const Bio = () => {
     }
   `)
 
-  const [author] = useJsonForm(data.dataJson, {
+  const [author] = useLocalJsonForm(data.dataJson, {
     label: "Author",
     fields,
   })