+++ packages/tinacms/src/plugins/fields/ImageFieldPlugin.tsx
@@ -20,6 +20,7 @@ import * as React from 'react'
 import { wrapFieldsWithMeta } from './wrapFieldWithMeta'
 import { InputProps, ImageUpload } from '@tinacms/fields'
 import { useCMS } from '../../react-tinacms'
+import { parse } from './textFormat'
 
 type FieldProps = any
 interface ImageProps {
@@ -59,4 +60,5 @@ export const ImageField = wrapFieldsWithMeta<InputProps, ImageProps>(props => {
 export default {
   name: 'image',
   Component: ImageField,
+  parse,
 }