+++ packages/tinacms/src/plugins/fields/MarkdownFieldPlugin.tsx
@@ -20,6 +20,7 @@ import { wrapFieldsWithMeta } from './wrapFieldWithMeta'
 import { Wysiwyg } from '@tinacms/fields'
 import styled from 'styled-components'
 import { InputCss } from '@tinacms/fields'
+import { parse } from './textFormat'
 
 const lightMediumGrey = `rgb(200, 200, 200)`
 const darkGrey = 'rgb(40, 40, 40)'
@@ -265,4 +266,5 @@ export const MarkdownField = wrapFieldsWithMeta(styled(Wysiwyg)`
 export default {
   name: 'markdown',
   Component: MarkdownField,
+  parse,
 }