+++ packages/tinacms/src/react-tinacms/index.ts
@@ -22,6 +22,7 @@ export * from './use-plugin'
 export * from './use-subscribable'
 export * from './use-watch-form-values'
 export * from './with-plugin'
+export * from './with-tina'
 
 export { Plugin } from '@tinacms/core'
 export { Form, FormOptions, Field } from '@tinacms/forms'