+++ packages/tinacms/src/tina-cms.ts
@@ -1,22 +1,22 @@
 /**
 
-Copyright 2019 Forestry.io Inc
+ Copyright 2019 Forestry.io Inc
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+ http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
 
-*/
+ */
 
-import { CMS, PluginType } from '@tinacms/core'
+import { CMS, CMSConfig, PluginType, Subscribable } from '@tinacms/core'
 import { FieldPlugin } from '@tinacms/form-builder'
 import { ScreenPlugin } from './plugins/screen-plugin'
 import TextFieldPlugin from './plugins/fields/TextFieldPlugin'
@@ -33,9 +33,22 @@ import GroupListFieldPlugin from './plugins/fields/GroupListFieldPlugin'
 import BlocksFieldPlugin from './plugins/fields/BlocksFieldPlugin'
 import { Form } from '@tinacms/forms'
 
+export declare type SidebarPosition = 'fixed' | 'float' | 'displace' | 'overlay'
+
+export interface TinaCMSConfig extends CMSConfig{
+  sidebar: {
+    hidden: boolean
+    position: SidebarPosition
+  }
+}
+
 export class TinaCMS extends CMS {
-  constructor() {
-    super()
+  sidebar: SidebarState
+
+  constructor({ sidebar, ...config}: TinaCMSConfig) {
+    super(config)
+
+    this.sidebar = new SidebarState(sidebar.position)
     this.fields.add(TextFieldPlugin)
     this.fields.add(TextareaFieldPlugin)
     this.fields.add(DateFieldPlugin)
@@ -60,5 +73,22 @@ export class TinaCMS extends CMS {
 
   get screens(): PluginType<ScreenPlugin> {
     return this.plugins.findOrCreateMap('screen')
+  }
+}
+
+export class SidebarState extends Subscribable {
+  constructor(public position: SidebarPosition) {
+    super()
+  }
+
+  private _isOpen: boolean = false
+
+  get isOpen() {
+    return this._isOpen
+  }
+
+  set isOpen(nextValue: boolean) {
+    this._isOpen = nextValue
+    this.notifiySubscribers()
   }
 }